Adicionando regra de negocio para o serviço de usuário

+++ Application/Services/Implementations/UserService.cs
@@ -17,7 +17,7 @@ namespace Application.Services.Implementations
         }
         public async Task<APIResponse<UserOutputDTO>> AllUsers()
         {
-            var users = await _userRepository.GetAll();
+            var users = await _userRepository.GetAllAsync();
 
             return Message.Response(
                 codeResponse: CodeEnum.OK,
@@ -29,6 +29,10 @@ namespace Application.Services.Implementations
 
         public async Task<APIResponse<UserOutputDTO>> CreateUser(UserCreateDTO userDTO)
         {
+            var usuarioExists = await _userRepository.ExistsAsync(p => p.Nickname == userDTO.NickName);
+
+            if (usuarioExists) return Message.Response<UserOutputDTO>(CodeEnum.BAD, Operation.USER_EXISTS, false, [], null);
+
             var usuarioEntity = userDTO.ToEntityInputInsert();
 
             usuarioEntity.EncryptPasswordEntity(HashPassword.CreatePasswordHash(userDTO.Password));
@@ -61,9 +65,14 @@ namespace Application.Services.Implementations
                 null);
         }
 
-        public async Task<APIResponse<UserOutputDTO>> DeleteUser(Guid userId)
+        public async Task<APIResponse<UserOutputDTO>> DeleteUser(UserDeleteDTO userDelete)
         {
-            var user = await _userRepository.GetUser(userId);
+            var user = await _userRepository.GetUserAsync(p => p.Nickname == userDelete.Nickname);
+
+            bool isPasswordValid = HashPassword.VerifyPasswordHash(userDelete.Password, user.Password!);
+
+            if (!isPasswordValid)
+                return Message.Response<UserOutputDTO>(CodeEnum.BAD, Operation.DELETE_FAILED, false, [], null);
 
             _userRepository.Delete(user);
 
@@ -79,7 +88,7 @@ namespace Application.Services.Implementations
 
         public async Task<APIResponse<UserOutputDTO>> User(Guid userId)
         {
-            var user = await _userRepository.GetUser(userId);
+            var user = await _userRepository.GetUserAsync(p => p.Id == userId);
 
             return Message.Response(
                 codeResponse: CodeEnum.OK,
